+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Changelog
 
 
+## 0.2.1 (2024-08-18)
+
+### Added
+
+- Testing for every transportation system
+- `Vehicle` object
+- `TransportationSystem.getVehicles()` which returns a list of `Vehicle` objects
+
+### Changed
+
+- Fixed `AttributeError: 'list' object has no attribute 'items'` error thrown in `TransportationSystem.getRoutes()`
+
+### Removed
+
+
 ## 0.2.0 (2024-07-31)
 
 ### Added